# feat: Implement memory search indexing for scalability (Issue #984)