--- a/security-audit-report.md
+++ b/security-audit-report.md
@@ -1,30 +1,20 @@
 # Security Audit Report
 
-Generated: 2025-09-17T20:07:13.879Z
-Duration: 17ms
+Generated: 2025-09-17T20:36:10.254Z
+Duration: 182ms
 
 ## Summary
 
-- **Total Findings**: 1
-- **Files Scanned**: 1
+- **Total Findings**: 0
+- **Files Scanned**: 132
 
 ### Findings by Severity
 
 - 🔴 **Critical**: 0
 - 🟠 **High**: 0
 - 🟡 **Medium**: 0
-- 🟢 **Low**: 1
+- 🟢 **Low**: 0
 - ℹ️ **Info**: 0
-
-## Detailed Findings
-
-### LOW (1)
-
-#### DMCP-SEC-006: Security operation without audit logging
-
-- **File**: `/var/folders/kj/45kjdq714853c8nlnsv7l0_r0000gn/T/security-audit-test-Qladfu/auth-handler.js`
-- **Confidence**: medium
-- **Remediation**: Add SecurityMonitor.logSecurityEvent() for audit trail
 
 ## Recommendations