--- a/src/elements/memories/index.ts
+++ b/src/elements/memories/index.ts
@@ -4,3 +4,4 @@
 
 export { Memory, MemoryMetadata, MemoryEntry, MemorySearchOptions } from './Memory.js';
 export { MemoryManager } from './MemoryManager.js';
+export { MemorySearchIndex, SearchQuery, SearchIndexConfig, SearchResult } from './MemorySearchIndex.js';