… .dollhousemcp/cache/collection-cache.json
… docs/development/SESSION_NOTES_2025_09_17_EVENING_INDEXING.md