@mickdarling 2025-09-17
## ✅ Security Audit Issue Fixed in commit 2503bdf

Fixed the LOW severity security audit finding.

[**View the fix here**](https://github.com/DollhouseMCP/mcp-server/pull/991/commits/2503bdf)

### Security Issue Fixed:

**DMCP-SEC-006: Security operation without audit logging**
- **Severity**: LOW  
- **File**: MemorySearchIndex.ts
- **Fix**: Added SecurityMonitor.logSecurityEvent() calls for all security operations

### Implementation:
- Added audit logging for Unicode validation failures
- Added audit logging for successful index builds  
- Added audit logging for search query validation
- All security operations now have proper audit trails

### Security Events Now Logged:
- `UNICODE_VALIDATION_FAILED` - When invalid Unicode is detected during indexing or search
- `MEMORY_INDEX_BUILT` - When index is successfully built (for monitoring)

## Summary Status

All review feedback has been addressed:
- ✅ Race condition protection (commit ce43ee7)
- ✅ Memory management & LRU eviction (commit ce43ee7)
- ✅ Enhanced error handling (commit ce43ee7)
- ✅ Index persistence/serialization (commit ce43ee7)
- ✅ Security audit logging (commit 2503bdf)

**Ready for final review and merge!** 🚀

### Note on Review Comments
It appears the reviewer may have been looking at an older version when commenting, as they didn't notice the race condition fixes (isBuilding flag, buildQueue) that were already implemented. All suggested improvements have been addressed in the commits above.

Co-Authored-By: Alex Sterling <alex.sterling@ai>