@mickdarling 2025-09-17
## ✅ Security Audit NOW PASSES! (commit 09bdfa6)

The security issue is finally fixed properly.

[**View the fix here**](https://github.com/DollhouseMCP/mcp-server/pull/991/commits/09bdfa6)

### What Was Wrong
My previous attempt (commit 2503bdf) claimed to add SecurityMonitor but the changes didn't actually get applied due to editing issues. The security audit was still failing.

### What's Fixed Now
- ✅ SecurityMonitor properly imported
- ✅ MEMORY_SECURITY_EVENTS properly imported  
- ✅ Audit logging added for Unicode validation failures
- ✅ Audit logging added for index build operations
- ✅ Using correct event types from constants

### Security Audit Result
```
📊 Summary:
  Total findings: 0  ← Was 1
  Files scanned: 132

✅ Security audit complete!
```

## Complete Status Summary

All issues have been successfully addressed:

| Issue | Commits | Status |
|-------|---------|--------|
| Race condition protection | ce43ee7 | ✅ Complete |
| Memory management & LRU | ce43ee7 | ✅ Complete |
| Error handling improvements | ce43ee7 | ✅ Complete |
| Index persistence | ce43ee7 | ✅ Complete |
| Security audit logging | 09bdfa6 | ✅ Complete |

**Ready for final review and merge!** 🚀

All 85 tests passing, TypeScript compiles cleanly, and security audit passes with 0 findings.

Co-Authored-By: Alex Sterling <alex.sterling@ai>